Move new wallet dialogs to separate file
