+++ android/app/src/main/res/layout/new_wallet_2.xml
